pyproject.toml
[project]
name = "banking-realtime-warehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=3.1.3",
    "apache-airflow-providers-amazon>=9.17.0",
    "apache-airflow-providers-snowflake>=6.6.1",
    "boto3>=1.41.3",
    "dbt-snowflake>=1.10.3",
    "faker>=38.2.0",
    "fastparquet>=2024.11.0",
    "kafka-python>=2.3.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
